everything in fclib.h, tests ok.
